OLMIS-3533: Added piority to reason type and used it to sort line items
The method sort line items by two dates fields but if there are two
line items (or more) with the same dates (in both fields) the order is
randomly and in some cases a DEBIT line item can be before a CREDIT line
item. This situation is dangerous when the DEBIT line item is the first in
the list because 'O - line_item_quantity' will get negative a value which
is not allowed.